feat: add parsing helper functions
